Added options form and advanced cell handling

+++ Miner2/MinerLogic/GameData.cs
@@ -12,10 +12,10 @@ namespace MinerLogic
         private GameField _gameField;
         private Options _options;
         private int _elapsedTime; //отображает затраченное время на форме
-        private int _minesLeft; //отображает количество оставшихся мин на форме
+        private short _minesLeft; //отображает количество оставшихся мин на форме
         private GameState _gameState; // состояние игры
 
-        public GameData(Options options, GameField gameField, int elapsedTime, int minesLeft, GameState gameState)
+        public GameData(Options options, GameField gameField, int elapsedTime, short minesLeft, GameState gameState)
         {
             _options = options;
             _gameField = gameField;
@@ -25,7 +25,7 @@ namespace MinerLogic
         }
 
         internal int ElapsedTime { get => _elapsedTime; set => _elapsedTime = value; }
-        internal int MinesLeft { get => _minesLeft; set => _minesLeft = value; }
+        internal short MinesLeft { get => _minesLeft; set => _minesLeft = value; }
         internal GameField GameField { get => _gameField; set => _gameField = value; }
         internal Options Options { get => _options; set => _options = value; }
         internal GameState GameState { get => _gameState; set => _gameState = value; }
